requirements.txt
Django
django-bootstrap-toolkit
django-registration-redux